fix: change return types of database-related functions to int for better error handling

+++ include/btree/page.h
@@ -3,7 +3,7 @@
 
 #include <stdio.h>
 
-unsigned short get_page_size(FILE *database_file);
+int get_page_size(FILE *database_file);
 int get_page_offset(int page_no);
 short get_page_header_size(FILE *database_file, int page_no);
